requirements.txt
torch --index-url https://download.pytorch.org/whl/cu121
numpy
scipy
matplotlib
h5py
neuraloperator
torchdiffeq
